refactor(persist): move AutomatePersistFailure from appevent/ to persist/
It is the only persist-outcome event in appevent/; siblings there are
FSM lifecycle events. Co-locating with PersistOutcome and ErrorCategory
matches its actual cohesion partners.

Still implements AppEvent and still fires through AutomateListener.
automatePersistFailure — only the declaration moves. No behavior change,
no public API change beyond the package path.
