+++ s2-automate/s2-automate-core/src/commonMain/kotlin/s2/automate/core/persist/AutomatePersistFailure.kt
@@ -1,13 +1,13 @@
-package s2.automate.core.appevent
+package s2.automate.core.persist
 
-import s2.automate.core.persist.ErrorCategory
+import s2.automate.core.appevent.AppEvent
 import s2.dsl.automate.S2Error
 
 /**
  * Notification when a PersistOutcome.Failure is produced by the persister.
  *
- * Symmetric to the success-path events ([AutomateTransitionEnded] /
- * [AutomateInitTransitionEnded]): one event per failed outcome so
+ * Symmetric to the success-path events (AutomateTransitionEnded /
+ * AutomateInitTransitionEnded): one event per failed outcome so
  * observers can implement dead-letter, alerting, or metrics.
  *
  * @param msgId the message id that triggered the transition
