+++ s2-automate/s2-automate-core/src/commonMain/kotlin/s2/automate/core/storing/S2AutomateStoringEvolverImpl.kt
@@ -9,11 +9,11 @@ import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.flow.flowOf
 import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.onEach
-import s2.automate.core.appevent.AutomatePersistFailure
 import s2.automate.core.appevent.listener.AutomateListener
 import s2.automate.core.appevent.publisher.AppEventPublisher
 import s2.automate.core.engine.S2AutomateEngine
 import s2.automate.core.engine.S2AutomateOutcomeEngine
+import s2.automate.core.persist.AutomatePersistFailure
 import s2.automate.core.persist.PersistOutcome
 import s2.dsl.automate.Evt
 import s2.dsl.automate.S2Automate
